requirements.txt
numpy~=2.1.0
matplotlib~=3.9.2
PyYAML~=6.0.2
pytest~=8.3.3